requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
schedule>=1.1.0
pytz>=2022.1
python-dateutil>=2.8.2